+++ pyopentsdb/conf.py
@@ -1,3 +1,23 @@
+from enum import Enum
+
+
+class StatsPointer(Enum):
+    STATS = 1
+    JVM = 2
+    QUERY = 3
+    REGION_CLIENTS = 4
+    THREADS = 5
+
+
+class ConfigPointer(Enum):
+    CONFIG = 1
+    FILTERS = 2
+
+
+class QueryPointer(Enum):
+    QUERY = 1
+
+
 BASIC_TSDB_QUERY = {
         'start': None,
         'end': None,
@@ -13,3 +33,6 @@ BASIC_TSDB_QUERY = {
         'useCalendar': False,
         'queries': list(),
     }
+
+
+