+++ pyopentsdb/config.py
@@ -1,34 +1,29 @@
 from pyopentsdb import utils
+from pyopentsdb.conf import ConfigPointer
 
 
-def tsdb_configuration(host, port, protocol, timeout):
-    url = api_url(host, port, protocol, pointer='CONFIG')
-    return utils.request_get(url, timeout)
+def tsdb_configuration(host, r_session, **kwargs):
+    return utils.request_get(api_url(host, pointer=ConfigPointer.CONFIG), r_session, **kwargs)
 
 
-def filters(host, port, protocol, timeout):
+def filters(host, r_session, **kwargs):
     """
     :param host: str
-    :param port: str
-    :param protocol: str
-    :param timeout: int/float/tuple
+    :param r_session: requests.Session
     :return: dict
     """
-    url = api_url(host, port, protocol, pointer='FILTERS')
-    return utils.request_get(url, timeout)
+    return utils.request_get(api_url(host, pointer=ConfigPointer.FILTERS), r_session, **kwargs)
 
 
-def api_url(host, port, protocol, pointer):
+def api_url(host, pointer):
     """
     Make api url to obtain configuration
 
     :param host: str
-    :param port: str
-    :param protocol: str
     :param pointer: str
     :return: str
     """
-    if pointer == 'CONFIG':
-        return '{}://{}:{}/api/config/'.format(protocol, host, port)
-    elif pointer == 'FILTERS':
-        return '{}://{}:{}/api/config/filters/'.format(protocol, host, port)
+    if pointer == ConfigPointer.CONFIG:
+        return '{}/api/config/'.format(host)
+    elif pointer == ConfigPointer.FILTERS:
+        return '{}/api/config/filters/'.format(host)