+++ pyopentsdb/stats.py
@@ -1,39 +1,33 @@
 from pyopentsdb import utils
+from pyopentsdb.conf import StatsPointer
 
 
-def stats(host, port, protocol, timeout):
+def stats(host, r_session, **kwargs):
     """
     This endpoint provides a list of statistics for the running TSD.
     Sub endpoints return details about other TSD components such as the JVM,
     thread states or storage client. All statistics are read only.
 
     :param host: str
-    :param port: str
-    :param protocol: str
-    :param timeout: int/float/tuple
+    :param r_session: requests.Session
     :return: dict
     """
-    url = api_url(host, port, protocol, pointer='STATS')
-    return utils.request_get(url, timeout)
+    return utils.request_get(api_url(host, pointer=StatsPointer.STATS), r_session, **kwargs)
 
 
-def jvm(host, port, protocol, timeout):
+def jvm(host, r_session, **kwargs):
     """
     The threads endpoint is used for debugging the TSD's JVM process and includes
     stats about the garbage collector, system load and memory usage. (v2.2)
 
     :param host: str
-    :param port: str
-    :param protocol: str
-    :param timeout: int/float/tuple
+    :param r_session: requests.Session
     :return: dict
     """
-
-    url = api_url(host, port, protocol, pointer='JVM')
-    return utils.request_get(url, timeout)
+    return utils.request_get(api_url(host, pointer=StatsPointer.JVM), r_session, **kwargs)
 
 
-def query(host, port, protocol, timeout):
+def query(host, r_session, **kwargs):
     """
     This endpoint can be used for tracking and troubleshooting queries executed
     against a TSD. It maintains an unbounded list of currently executing
@@ -43,63 +37,52 @@ def query(host, port, protocol, timeout):
     if thrown. (v2.2)
 
     :param host: str
-    :param port: str
-    :param protocol: str
-    :param timeout: int/float/tuple
+    :param r_session: requests.Session
     :return: json
     """
-    url = api_url(host, port, protocol, pointer='QUERY')
-    return utils.request_get(url, timeout)
+    return utils.request_get(api_url(host, pointer=StatsPointer.QUERY), r_session, **kwargs)
 
 
-def region_clients(host, port, protocol, timeout):
+def region_clients(host, r_session, **kwargs):
     """
     Returns information about the various HBase region server clients in AsyncHBase.
     This helps to identify issues with a particular region server. (v2.2)
 
     :param host: str
-    :param port: str
-    :param protocol: str
-    :param timeout: int/float/tuple
+    :param r_session: requests.Session
     :return: json
     """
-    url = api_url(host, port, protocol, pointer='REGION_CLIENTS')
-    return utils.request_get(url, timeout)
+    return utils.request_get(api_url(host, pointer=StatsPointer.REGION_CLIENTS), r_session, **kwargs)
 
 
-def threads(host, port, protocol, timeout):
+def threads(host, r_session, **kwargs):
     """
     The threads endpoint is used for debugging the TSD and providing insight
     into the state and execution of various threads without having to resort
     to a JStack trace. (v2.2)
 
     :param host: str
-    :param port: str
-    :param protocol: str
-    :param timeout: int/float/tuple
+    :param r_session: requests.Session
     :return: json
     """
-    url = api_url(host, port, protocol, pointer='THREADS')
-    return utils.request_get(url, timeout)
+    return utils.request_get(api_url(host, pointer=StatsPointer.THREADS), r_session, **kwargs)
 
 
-def api_url(host, port, protocol, pointer):
+def api_url(host, pointer):
     """
     Make api url to obtain configuration
 
     :param host: str
-    :param port: str
-    :param protocol: str
     :param pointer: str
     :return: str
-    # """
-    if pointer == 'STATS':
-        return '{}://{}:{}/api/stats/'.format(protocol, host, port)
-    elif pointer == 'JVM':
-        return '{}://{}:{}/api/stats/jvm/'.format(protocol, host, port)
-    elif pointer == 'QUERY':
-        return '{}://{}:{}/api/stats/query/'.format(protocol, host, port)
-    elif pointer == 'REGION_CLIENTS':
-        return '{}://{}:{}/api/stats/region_clients/'.format(protocol, host, port)
-    elif pointer == 'THREADS':
-        return '{}://{}:{}/api/stats/threads/'.format(protocol, host, port)
+    """
+    if pointer == StatsPointer.STATS:
+        return '{}/api/stats/'.format(host)
+    elif pointer == StatsPointer.JVM:
+        return '{}/api/stats/jvm/'.format(host)
+    elif pointer == StatsPointer.QUERY:
+        return '{}/api/stats/query/'.format(host)
+    elif pointer == StatsPointer.REGION_CLIENTS:
+        return '{}/api/stats/region_clients/'.format(host)
+    elif pointer == StatsPointer.THREADS:
+        return '{}/api/stats/threads/'.format(host)