+++ pyopentsdb/suggest.py
@@ -4,45 +4,44 @@ from pyopentsdb import errors
 from pyopentsdb import utils
 
 
-def suggest(host, port, protocol, timeout, **kwargs):
-    url = api_url(host, port, protocol)
+def suggest(host, r_session, **kwargs):
     try:
-        t = kwargs['type']
+        t = kwargs.pop('type')
     except KeyError:
         raise errors.MissingArgumentError("'type' is a required argument")
-
-    params = dict()
-    params.update(
+    kwargs.update(
         {
-            'type': t,
-            'q': kwargs.get('q', ''),
-            'max': int(kwargs.get('max', 25))
-        }
-    )
+            "data":
+                {
+                    'type': t,
+                    'q': kwargs.pop('q', ''),
+                    'max': int(kwargs.pop('max', 25))
+                }
+            }
+        )
+    return utils.request_post(api_url(host), r_session, **kwargs)
 
-    return utils.request_post(url, params, timeout)
 
+def metrics(host, r_session, **kwargs):
+    kwargs.pop("type", None)
+    regexp = kwargs.pop('regexp', None)
 
-def metrics(host, port, protocol, timeout, **kwargs):
     kwargs.update({
         'type': 'metrics',
-        'q': kwargs.get('q', ''),
-        'max': kwargs.get('max', 10000),
+        'q': kwargs.pop('q', ''),
+        'max': kwargs.pop('max', 10000),
     })
 
-    metric_list = suggest(host, port, protocol, timeout, **kwargs)
-    regexp = kwargs.get('regexp')
+    metric_list = suggest(host, r_session, **kwargs)
     if regexp:
         metric_list = [metric for metric in metric_list if re.search(re.compile(regexp), str(metric), flags=0)]
     return metric_list
 
 
-def api_url(host, port, protocol):
+def api_url(host):
     """
     Make api url to obtain configuration
     :param host: str
-    :param port: str
-    :param protocol: str
     :return: str
     """
-    return '{}://{}:{}/api/suggest/'.format(protocol, host, port)
+    return '{}/api/suggest/'.format(host)