+++ tests/test_version.py
@@ -20,17 +20,15 @@ class TsdbVersionTestCase(unittest.TestCase):
     }
 
     def setUp(self):
-        self._host = 'localhost'
-        self._port = 5896
-        self._protocol = 'mockhttp'
+        self._host = 'mockhttp://localhost:5896/'
 
-        self._c = tsdb.tsdb_connection(self._host, self._port, protocol=self._protocol)
+        self._c = tsdb.tsdb_connection(self._host)
 
-    @mock.patch('requests.get', side_effect=get_mock_requests_get(None))
+    @mock.patch('requests.Session.get', side_effect=get_mock_requests_get(None))
     def test_url(self, _):
         GeneralUrlTestCase.test_url(self, "/api/version/", "version")
 
-    @mock.patch('requests.get', side_effect=get_mock_requests_get(__TEST_VERSION__))
+    @mock.patch('requests.Session.get', side_effect=get_mock_requests_get(__TEST_VERSION__))
     def test_version(self, _):
         return_value = self._c.version()
         self.assertEqual(sorted(return_value), sorted(TsdbVersionTestCase.__TEST_VERSION__))