+++ tests/testutils.py
@@ -6,6 +6,7 @@ from pyopentsdb import tsdb
 ADHOC_HOSTS = ['localhost', 'myqeb.com', 'web.kim.com']
 ADHOC_PORTS = ['201', 22222, 693]
 ADHOC_PROTOCOLS = ['http', 'https', 'ftp']
+ADHOC_HOSTS = ["{}://{}:{}/".format(a, b, c) for a, b, c in zip(ADHOC_PROTOCOLS, ADHOC_HOSTS, ADHOC_PORTS)]
 
 
 class MockRequests(object):
@@ -96,11 +97,19 @@ def mock_unexpected_error_get(url, *args, **kwargs):
     raise Exception("Unexpected error")
 
 
+def get_mock_utils_requests_post(requests_kwargs):
+    def mock_utils_request_post(url, r_session, **kwargs):
+        _ = kwargs.pop("data")
+        not_poped = [kwarg_key for kwarg_key in kwargs if kwarg_key not in requests_kwargs]
+        return dict(status=False) if not_poped else dict(status=True)
+    return mock_utils_request_post
+
+
 class GeneralUrlTestCase(object):
     def test_url(self, endpoint, exec_fn, **kwargs):
-        for a, b, c in zip(ADHOC_HOSTS, ADHOC_PORTS, ADHOC_PROTOCOLS):
-            expected_url = '{}://{}:{}{}'.format(c, a, b, endpoint)
-            c = tsdb.tsdb_connection(a, b, protocol=c)
+        for host in ADHOC_HOSTS:
+            expected_url = '{}{}'.format(host[:-1], endpoint)
+            c = tsdb.tsdb_connection(host)
             _exec_fn = getattr(c, exec_fn)
             mock_return_value = _exec_fn(**kwargs)
             self.assertTrue(mock_return_value[-1] == expected_url)